Moved Charcters into its own page
